refactor: move series preferences to new module
Move the SeriesPreferences into a new core:preferences module. This module will be needed to share
any preference between modules and the settings module.
