+++ app/build.gradle
@@ -53,6 +53,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core:preferences")
     implementation project(":features:discover")
     implementation project(":features:login")
     implementation project(":features:profile")
