+++ app/src/androidTest/java/com/chesire/nekome/helpers/SeriesPreferencesExtensions.kt
@@ -1,7 +1,7 @@
 package com.chesire.nekome.helpers
 
-import com.chesire.nekome.app.series.SeriesPreferences
 import com.chesire.nekome.core.flags.SortOption
+import com.chesire.nekome.core.preferences.SeriesPreferences
 
 /**
  * Resets the [SeriesPreferences] back to a default state.
