+++ core/preferences/build.gradle
@@ -1,0 +1,34 @@
+plugins {
+    id "com.android.library"
+    id "kotlin-android"
+    id "dagger.hilt.android.plugin"
+    id "kotlin-kapt"
+}
+
+android {
+    namespace 'com.chesire.nekome.core.preferences'
+    compileSdk sdk_version
+
+    defaultConfig {
+        minSdk 21
+        targetSdk sdk_version
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
+        }
+    }
+}
+
+dependencies {
+    implementation project(":libraries:core")
+
+    implementation libs.androidx.datastore.preferences
+    implementation libs.google.hilt.android
+    implementation libs.squareup.moshi
+    kapt libs.google.hilt.android.compiler
+}
