+++ features/series/build.gradle
@@ -35,6 +35,7 @@ android {
 
 dependencies {
     implementation project(":core:compose")
+    implementation project(":core:preferences")
     implementation project(":libraries:core")
     implementation project(":libraries:datasource:series")
