+++ features/series/src/main/java/com/chesire/nekome/app/series/collection/core/ShouldRateSeriesUseCase.kt
@@ -1,6 +1,6 @@
 package com.chesire.nekome.app.series.collection.core
 
-import com.chesire.nekome.app.series.SeriesPreferences
+import com.chesire.nekome.core.preferences.SeriesPreferences
 import com.chesire.nekome.datasource.series.SeriesRepository
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
