+++ features/series/src/test/java/com/chesire/nekome/app/series/collection/core/FilterSeriesUseCaseTest.kt
@@ -2,9 +2,9 @@
 
 package com.chesire.nekome.app.series.collection.core
 
-import com.chesire.nekome.app.series.SeriesPreferences
 import com.chesire.nekome.core.flags.SeriesType
 import com.chesire.nekome.core.flags.UserSeriesStatus
+import com.chesire.nekome.core.preferences.SeriesPreferences
 import com.chesire.nekome.testing.createSeriesDomain
 import io.mockk.clearAllMocks
 import io.mockk.every
