+++ features/series/src/test/java/com/chesire/nekome/app/series/collection/core/UpdateSortUseCaseTest.kt
@@ -2,8 +2,8 @@
 
 package com.chesire.nekome.app.series.collection.core
 
-import com.chesire.nekome.app.series.SeriesPreferences
 import com.chesire.nekome.core.flags.SortOption
+import com.chesire.nekome.core.preferences.SeriesPreferences
 import io.mockk.clearAllMocks
 import io.mockk.coEvery
 import io.mockk.just
